backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.34.0
chromadb==0.5.15
pypdf==4.3.1
python-dotenv==1.0.1
pydantic==2.9.2
sentence-transformers==3.3.1
